requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Email & Scheduling
apscheduler==3.10.4
jinja2==3.1.2
python-dateutil==2.8.2
lunarcalendar==0.0.9

# Google API (OAuth 2.0)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
